Renaming files and implementing Remove Lecture event
